tasks/parallel-web-scraping-aws/processor/requirements.txt
boto3
beautifulsoup4
google-auth
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
python-dotenv